refactor: 사용자 정보 로딩과 관련된 상태 관리 로직을 훅으로 분리
